+++ src/constants/Extension.ts
@@ -25,6 +25,8 @@ export const COMMAND_NAME = {
   registerFolder: getCommandName("registerFolder"),
   unregisterFolder: getCommandName("unregisterFolder"),
   createContent: getCommandName("createContent"),
+  createByContentType: getCommandName("createByContentType"),
+  createByTemplate: getCommandName("createByTemplate"),
   createTemplate: getCommandName("createTemplate"),
   collapseSections: getCommandName("collapseSections"),
   preview: getCommandName("preview"),